refactor : Mocks alias 추가, 경로 수정
