+++ src/__mocks__/images/squidgame-spotdetail.jpg
