+++ src/__mocks__/images/squidgame-spotdetail2.jpg
