+++ src/__mocks__/images/squidgame-spotdetail3.png
