+++ src/__mocks__/images/squidgame-spotdetail4.png
