+++ tsconfig.app.json
@@ -23,7 +23,8 @@
       "@/entities/*": ["src/entities/*"],
       "@/features/*": ["src/features/*"],
       "@/pages/*": ["src/pages/*"],
-      "@/app/*": ["src/app/*"]
+      "@/app/*": ["src/app/*"],
+      "@/__mocks__/*": ["src/__mocks__/*"]
     },
 
     /* Linting */
